config V show.

+++ Code/CubeMX/TouchGFX/gui/include/gui/model/Model.hpp
@@ -31,7 +31,7 @@ class Model {
     ModelListener *modelListener;
     int8_t ctrl_keyValue;
     int8_t keyValue;
-    int8_t voltage;
+    double voltage;
     int8_t rsL_x;
     int8_t rsL_y;
     int8_t rsR_x;
